requirements.txt
redis==6.0.0
langgraph==0.4.3
langchain==0.3.21
langchain-community==0.3.20
python-telegram-bot==22.0
loguru==0.7.3
duckduckgo-search==7.5.4
beautifulsoup4==4.13.3
requests==2.32.3
chromadb==1.0.8
torch==2.7.0 
torchvision==0.22.0